tests/wpt/web-platform-tests/notifications/constructor-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notification constructor (invalid)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    if (Notification.permission != "granted") {
        this.force_timeout()
        this.set_status(this.NOTRUN, "You must allow notifications for this"
            + " origin before running this test.")
    }
    assert_throws_js(TypeError, function() {
        new Notification()
    })
}, "Called the notification constructor with no arguments.")
</script>
